notifier/requirements.txt
twilio==6.13.0
requests==2.18.4
pyzmq==18.0.1
